Helicopter enemy now shoots projectiles at the player. Projectiles arc with gravity to emulate water stream. When helicopter cannot see player it will move closer to player till it can see player

+++ Assets/Scripts/PlayerHealth.cs
@@ -3,8 +3,8 @@ using UnityEngine.SceneManagement;
 
 public class PlayerHealth : MonoBehaviour
 {
-    public int maxHealth = 100;
-    private int currentHealth;
+    public float maxHealth = 100f;
+    private float currentHealth;
     public GameObject Canvas;
     private Canvas UImanager;
     public bool isAlive = true;
@@ -30,7 +30,7 @@ public class PlayerHealth : MonoBehaviour
     }
 
     // function that other scripts can call to Deal Damage
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
 
         if (!isAlive) return;
